Asynchonous queuing for SNS/SQS publish
Added a thread pool to Propono so that SNS publish is handled asynchronously. A
future enhancment may make this configurable, but for now this is the only
request mode.

A call to publish to SNS returns a future, so publishers desiring a blocking
implementation can block on the future. The future's value is the AWS response
return by Fog.

	response = publisher.send(:publish_via_sns)
	response.value # (block on value)